pyproject.toml
[project]
name = "freshbooks-tools"
version = "0.1.0"
description = "CLI tool for querying FreshBooks time entries and invoices"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "textual>=0.89",
    "pydantic>=2.9",
    "platformdirs>=4.3",
    "pyyaml>=6.0",
]

[project.scripts]
fb = "freshbooks_tools.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/freshbooks_tools"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
